requirements.txt
ipykernel
opendatasets
streamlit
imblearn
scikit-learn
matplotlib
seaborn
pandas
numpy
python-dotenv
PyYAML
evidently
dill

-e .